Extract bool input to separate component
